+++ e2e-tests/helpers/openai-compatible.ts
