+++ e2e-tests/specs/active-note-extraction.spec.ts
