+++ e2e-tests/specs/folder-extraction.spec.ts
