+++ e2e-tests/specs/old-installer-version.spec.ts
