+++ e2e-tests/specs/services/mistral.spec.ts
