+++ e2e-tests/specs/services/openai-compatible-settings.spec.ts
