+++ e2e-tests/specs/services/openai-compatible.spec.ts
