+++ e2e-tests/test-vault/.obsidian/community-plugins.json
