+++ e2e-tests/test-vault/attachments/no_text.pdf
