+++ e2e-tests/test-vault/attachments/sample.png
